rewritten all tests to unittest2; get rid of nimskull

+++ tests/ops/test_openat.nim
@@ -1,14 +1,20 @@
-import balls
+import unittest2
 import nimuring
 import posix
 import os
 
-var q = newQueue(1, {})
+template withTempFile(body: untyped) =
+  let path {.inject.} = getTempDir() / "test_io_uring_openat"
+  try:
+    body
+  finally:
+    if fileExists(path):
+      removeFile(path)
 
-const path = getTempDir() / "test_io_uring_openat";
-
-q.openat(0, 0, path, O_CLOEXEC or O_RDWR or O_CREAT)
-q.submit(1)
-
-check fileExists(path)
-removeFile(path)
+suite "openat operation":
+  test "openat creates file":
+    withTempFile:
+      var q = newQueue(1, {})
+      q.openat(0, 0, path, O_CLOEXEC or O_RDWR or O_CREAT)
+      q.submit(1)
+      check fileExists(path)
